Quick README annotation concerning images.
The images in my README file were appearing correctly in the preview for my VS Code, but were not appearing at all when viewing the README on GitHub. I do not know why this is, so I made a note of it at the end of the README file, which was the only change I made for this commit.